skin_disease_detection/requirements.txt
Flask
tensorflow
numpy
Pillow
pandas
scikit-learn
